+++ package.json
@@ -49,6 +49,9 @@
   "devDependencies": {
     "@commitlint/cli": "^19.0.3",
     "@commitlint/config-conventional": "^19.0.3",
+    "@types/fs-extra": "^11.0.4",
+    "@types/lodash.get": "^4.4.9",
+    "@types/lodash.trimstart": "^4.5.9",
     "eslint": "^8.39.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-prettier": "^5.0.0",
@@ -56,7 +59,8 @@
     "lint-staged": "^15.0.2",
     "markdown-toc": "^1.2.0",
     "prettier": "^3.0.1",
-    "tap": "^16.3.4"
+    "tap": "^16.3.4",
+    "typescript": "^5.4.5"
   },
   "lint-staged": {
     "*.{js,jsx}": "eslint --cache --fix"
