+++ src/utils/do-not-edit-text.ts
@@ -1,9 +1,8 @@
-const doNotEditText = `/* eslint-disable */
+export const doNotEditText = `/* eslint-disable */
 /**
  * This file was automatically generated by openapi-transformer-toolkit CLI/methods.
  * DO NOT MODIFY IT BY HAND. Instead, modify the source OpenAPI file,
  * and run openapi-transformer-toolkit CLI/methods to regenerate this file.
  */
- 
+
  `
-export { doNotEditText }
