--- src/utils/openapi-schema-to-json-schema-wrapper.cjs
@@ -1,2 +1,0 @@
-const { fromSchema } = require('@openapi-contrib/openapi-schema-to-json-schema')
-module.exports = { fromSchema }
