+++ src/utils/openapi-schema-to-json-schema-wrapper.ts
@@ -1,0 +1,2 @@
+export { fromSchema } from '@openapi-contrib/openapi-schema-to-json-schema';
+
