+++ src/utils/paths.ts
@@ -1,0 +1,20 @@
+import { join as joinDesm } from 'desm'
+import { join, resolve } from 'node:path'
+import { cwd } from 'node:process'
+
+const workingDirectory = cwd()
+const packageRoot = joinDesm(import.meta.url, '..', '..')
+
+const resolvePath = (basePath: string, ...pathParts: string[]): string => {
+  try {
+    return resolve(join(basePath, ...pathParts))
+  } catch {
+    return ''
+  }
+}
+
+export const resolveFromPackageRoot = (...pathParts: string[]) =>
+  resolvePath(packageRoot, ...pathParts)
+
+export const resolveFromWorkingDirectory = (...pathParts: string[]) =>
+  resolvePath(workingDirectory, ...pathParts)
